Support "q" and "scope" query parameters
This adds support for "q" and "scope" parameters in the URL:
- `q`: This value is put into the search field on the page for filtering rows that match the provided text
- `scope`: The value corresponds to the "All, Release, Betas" radio buttons

Additionally, the logic of the page has changed slightly such that the URL (and not the input values) are now the source of truth: updating the input fields changes the URL, and then the URL is used to filter the rows. When the page loads, any existing URL values are placed into the input fields so that state remains consistent.

Kudos to @kj800x for the idea: https://twitter.com/kevinismahname/status/1595117874653573120